+++ src/lib/wavelettransform.cpp
@@ -21,19 +21,19 @@ WaveletTransform::WaveletTransform(std::shared_ptr<Wavelet> wavelet, int numLeve
 WaveletTransform::~WaveletTransform() {
 }
 
-void WaveletTransform::forward1d(VectorXd& signal) {
+void WaveletTransform::forward1d(VectorXf& signal) {
 	// signal must be 2^numlevels long
 	if (signal.size() % (1 << numLevels) != 0)
 		throw std::runtime_error("WaveletTransform::forward1d: signal must be 2^numLevels long!");
 
 	size_t size = signal.size();
 	for (int i = 0; i < numLevels; ++i) {
-		wavelet->forward(ArrayRef<double>(signal.data(), size));
+		wavelet->forward(ArrayRef<float>(signal.data(), size));
 		size /= 2;
 	}
 }
 
-void WaveletTransform::inverse1d(VectorXd& dwt) {
+void WaveletTransform::inverse1d(VectorXf& dwt) {
 	// input must be 2^numlevels long
 	if (dwt.size() % (1 << numLevels) != 0)
 		throw std::runtime_error("WaveletTransform::inverse1d: input must be 2^numLevels long!");
@@ -41,25 +41,25 @@ void WaveletTransform::inverse1d(VectorXd& dwt) {
 	// get second highest level size
 	size_t size = dwt.size() / (1 << (numLevels - 1));
 	for (int i = 0; i < numLevels; ++i) {
-		wavelet->inverse(ArrayRef<double>(dwt.data(), size));
+		wavelet->inverse(ArrayRef<float>(dwt.data(), size));
 		size *= 2;
 	}
 }
 
 void WaveletTransform::forward2d(cv::Mat& signal) {
-	assert(signal.type() == CV_64F);
+	assert(signal.type() == CV_32F);
 
 	cv::Mat roi(signal, cv::Rect(0, 0, signal.cols, signal.rows));
 	for (int i = 0; i < numLevels; ++i) {
 		// transform rows
 		for (int i = 0; i < roi.rows; ++i) {
-			ArrayRef<double> rowPtr(roi.ptr<double>(i), roi.cols);
+			ArrayRef<float> rowPtr(roi.ptr<float>(i), roi.cols);
 			wavelet->forward(rowPtr);
 		}
 		// transpose matrix so col transformation can be done in rows
 		cv::Mat transposed = roi.t();
 		for (int i = 0; i < transposed.rows; ++i) {
-			ArrayRef<double> rowPtr(transposed.ptr<double>(i), transposed.cols);
+			ArrayRef<float> rowPtr(transposed.ptr<float>(i), transposed.cols);
 			wavelet->forward(rowPtr);
 		}
 		// copy transposed to roi
@@ -71,20 +71,20 @@ void WaveletTransform::forward2d(cv::Mat& signal) {
 }
 
 void WaveletTransform::inverse2d(cv::Mat& dwt) {
-	assert(dwt.type() == CV_64F);
+	assert(dwt.type() == CV_32F);
 
 	size_t factor = 1 << (numLevels - 1);
 	cv::Mat roi(dwt, cv::Rect(cv::Point(0, 0), cv::Size(dwt.cols / factor, dwt.rows / factor)));
 	for (int i = 0; i < numLevels; ++i) {
 		// transform rows
 		for (int i = 0; i < roi.rows; ++i) {
-			ArrayRef<double> rowPtr(roi.ptr<double>(i), roi.cols);
+			ArrayRef<float> rowPtr(roi.ptr<float>(i), roi.cols);
 			wavelet->inverse(rowPtr);
 		}
 		// transpose matrix so col transformation can be done in rows
 		cv::Mat transposed = roi.t();
 		for (int i = 0; i < transposed.rows; ++i) {
-			ArrayRef<double> rowPtr(transposed.ptr<double>(i), transposed.cols);
+			ArrayRef<float> rowPtr(transposed.ptr<float>(i), transposed.cols);
 			wavelet->inverse(rowPtr);
 		}
 		// copy transposed to roi
